Avoid Polyfill supply chain attack ----------------------------------
The new owner of the polyfill.io domain is injecting JavaScript code to
serve uninteded content (commercials) to end users[1]. It could also be used
for other malicious campaigns, to inject targeted attack code.

To reduce the risk we're changing the polyfill to load from a "good copy"
made by CloudFlare[2] back in February 2024.

TODO: Reconsider the use of the polyfill library, is it still needed?

References:
[1] https://sansec.io/research/polyfill-supply-chain-attack
[2] https://blog.cloudflare.com/automatically-replacing-polyfill-io-links-with-cloudflares-mirror-for-a-safer-internet